pyproject.toml
[project]
name = "chat-server"
version = "0.1.0"
description = ""
authors = [
    {name = "jarvis",email = "jarvis@innosonian.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels[daphne] (>=4.2.0,<5.0.0)",
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytz (>=2025.1,<2026.0)",
    "pymongo (>=4.11,<5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.1"

